Move `ClearCache` command
